backend/requirements.txt
pandas==2.1.4
python-dotenv==1.0.0
supabase==1.0.3
python-jose==3.3.0
openai>=1.0.0 